Extract web_link statistics into concern
