+++ app/models/ahoy/visit.rb
@@ -1,6 +1,6 @@
 class Ahoy::Visit < ApplicationRecord
   self.table_name = "ahoy_visits"
 
-  has_many :events, class_name: "Ahoy::Event"
+  has_many :events, class_name: "Ahoy::Event", dependent: :destroy
   belongs_to :user, optional: true
 end
