+++ app/models/web_link.rb
@@ -1,4 +1,6 @@
 class WebLink < ApplicationRecord
+  include Statistics
+
   belongs_to :user, counter_cache: true
 
   before_validation :generate_random_uid, on: :create
@@ -7,27 +9,7 @@ class WebLink < ApplicationRecord
 
   validates :original_url, format: { with: VALID_URL_REGEX }
 
-  def browser_properties
-    joined_visits.group(:browser).count
-  end
-
-  def device_properties
-    joined_visits.group(:device_type).count
-  end
-
-  def created_at_properties
-    joined_visits.group_by_day(:started_at).count
-  end
-
-  def visit_count
-    Ahoy::Event.where(properties: { title: uid }).count
-  end
-
   private
-
-  def joined_visits
-    Ahoy::Visit.joins(:events).where(events: { properties: { title: uid } })
-  end
 
   def generate_random_uid
     generate_token(:uid)
